+++ cors/accepted_domains.ini
@@ -7,15 +7,18 @@
 
 
 # Examples of valid domains
-www.example.com
-www.example.com/forward-slash/
-www.*.gov
-*.mil
+# www.example.com
+# www.example.com/forward-slash/
+# www.*.gov
+# *.mil
 
 # Examples of invalid domains that are sanitized and still generate valid regex expressions.
-www.special_chars.com/`~!@#$%^&+=
-www.example.com\back-slash\
+# www.special_chars.com/`~!@#$%^&+=
+# www.example.com\back-slash\
 
 # An example of a raw regular expression used for CORS which accepts domains containing 'google'
 # that have a 3 letter suffix '.XYZ', i.e. '.com', '.org', etc.
-!^.*google.*.\.[\w]{3}$
+#!^.*google.*.\.[\w]{3}$
+
+# Allow all requests for testing - REMOVE THIS BEFORE DEPLOYING TO PRODUCTION
+*
